Replace list(set()) trick with actual remove_duplicates function
Sets only allow hashable objects, allowing for a more efficient
comparison and management of itself. Dict objects aren't
hashable, therefore can't be turned into set items. We replace that
trick with an actual function that reconstructs the list but only adds
items not already present. This function uses "obj not in list" notation
which uses `__eq__` and not `__hash__` interface, which `dict`
implements.

Included is a the resulting `generated-sources`.py which show the
resulting generated sources.