pip: handle spaces in package names
If you want to generate a manifest for diffoscope, including its
optional distro_detection and cmdline components. This invocation
correctly fetches the list of necessary packages:

    flatpak-pip-generator 'diffoscope[distro_detection, cmdline]'

But in the resulting file, the 'build-commands' line is not properly
quoted, so 'diffoscope[distro_detection,' and 'cmdline]' are passed as
separate arguments to 'pip3 install', and the build fails.

Fix this by shell-quoting arguments using shlex.quote().