go: make subprocess calls python <3.7 compatible
The code assumes Python 3.7 which is a tough requirement for some, e.g.
Ubuntu doesn't ship that version in all of its currently supported
versions.

This patch tries to mimick the intended behaviour, i.e. make it return
the decoded stdout. And keep failing if the command does not exit
cleanly.