Add the ability to use VSC packages
This change adds the ability for developers to provide VCS type packages
to `flatpak-pip-generator`.

> flatpak-pip-generator "git+https://github.com/user/repo#egg=name"

The changes will parse the VSC package name and ensure it is handed
throughout the rest of the process. The reason for this change is to
allow developers to build a flatpak using packages that may not, nor
ever be, on PyPi.

Signed-off-by: Kevin Carter <kecarter@redhat.com>