pip: Refactor the way packages are downloaded
Instead of downloading packages multiple times, do a single run of
`pip download -r requirements_file`, while the sources list for each
package is just a duplicate, the actual number of downloaded modules is
smaller as now `flatpak-builder` only downloads a single
version/extension.

Previously, if a package in the requirements file specified a version,
it could be ignored if a previous module has it as a dependency. Now
pip takes care of making sure version numbers are enforced.

Additionally, now it is possible to enfoce hashes via `pip download --require-hashes`
if `--hash` is used in requirements file.