pip: Install vcs modules first
Since vcs modules can only be installed with `pip install .`, if a
normal packages requires it, it won't be able to find it even if it is
in the sources list. Additionally, `flatpak-builder` will download `git`
sources as many times as they appear in the manifest and this reduces
the appearances of vcs sources, by listing them only for vcs modules.