pip: Manually downloads arch agnostic sources
Arch dependent packages are removed and replaced with their
correspondent arch agnostic versions. This removes the need for all
calls of `pip download --no-binary :all`.

Additionally, the only way duplicated files could be found is if there
is a vcs package which also happens to be a dependency of other package
in the list. The logic could be simplified further.