pip: Prettify output
Different sections of the script are separated in the same fashion of
`flatpak-builder`. It is clear what are the specific versions downloaded
and redownloaded.

Warn the user if his python version does not match the one in
org.freedesktop.Platform, some packages like `qdarkstyle` are specific
to the minor version of python and `flatpak-pip-generator`
generates useless output.