cargo: maintain compatibility with Python 3.6
The Python documentation states that "text" is a more readable alias for
"universinal_newlines". It was introduced with Python 3.7.
Ubuntu 18.04, however, run Python 3.6, is still supported, and doesn't
have an upgrade path yet.
This change brings compatibility with said Ubuntu version, but will
also benefit other Python 3.6 users.

The first version of this patch replaced "text"
with "universal_newlines" but this was thought to be less
readable than using decode().
Once Python 3.6 support is to be dropped, "text=True" can be used
for Popen.